feat: add local SSH client
APC Network Management Card V3 with the latest firmware uses a SSH Host Key algorithm that the Crestron SSH implmentation does not support. In order to use SSH with APC devices that use this combination, the plugin needs to use the full Renci SSH.NET implementation in order to allow the use of the correct host key algorithm.